requirements.txt
# Core dependencies only
requests>=2.31.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Audio processing
numpy>=1.24.0
pyaudio>=0.2.11
wave>=0.0.2
